Bracket Randomisation Improvement
Brackets can now encapsulate more than 2 cells at a time, having groups of 3 or more cells in parallel together. However, this means that large groups of cells are bundled together in parallel more often due to the randomisation of where the brackets are placed. For later, the probability distribution and number of brackets could be tweaked in order to generate more desirable configurations.