Partition function working, with minor caveat
Upon initial testing, the partition function seems to work as expected, however sometimes it generates two new rectangles as a result of splitting the grid. This should be fine for our applications later, but added in a warning clause just in case.

Rather than filtering out invalid stops, simply using the continue keyword allows the right kind of rectangular partitions to be generated.