Created make_block function, change_cell_interconnection function and continued building solar_module
The make_block function creates an all series connection given a list of cell ids that define the block. The change_cell_interconnection and change_connection method takes a circuit object and returns a new circuit with a specified format string or a random one created by generate_string.

The solar_module class contains attributes for each circuit corresponding to each block, including the format strings for each block and output nodes in order for plotting and power evaluation later.

The blocks now need to be connected together in series and parallel to create the second level of hierarchy.