Deleted change_cell_interconnection, created solar_block although bugged.
Doesn't seem to produce the right IV curves. Solar_block uses the undocumented copy_to function to copy the netlist of a circuit to another netlist. Perhaps a better idea is to concatenate the formatted strings together and then use interconnection function to build the entire module, rather than worrying about nested subcircuits as PySpice seems to dislike it.