solar_module functional with minor bugs
solar_module seems to be working with cell interconnection being changed randomly, but requires further testing. Adjacency flag takes a long time on large partitions. Added various testing cases. block_interconnection based on interconnection, but with a confusing pair of node trackers in order to ensure that the correct nodes are connected when blocks are in parallel or series.
TO DO: Test solar_module when all connections are changed, and on different sized grids.
Also, expand interconnection_testing to now use solar_module.py.